Add NetworkPromiseSpec protocol to reuse test code
